Update info panel to change person dispalyed
Previous commit was only displaying a test person, current implementation takes the person at the first index of the list to display on start up.
Commands such as `add`, `edit`, `find` and `pay` will cause the display to change to the person they are referring to.

Note that cases where there might be no contacts in the list has not been taken into account yet.