Update import functionality
Import is now able to take in filepaths with paths consisting of directories with spaces in their name.
Update tests to adjust to the updated functionality.

Adjust naming of tests according to convention.
Remove multiple_file_path test as function is now unable to distinguish between as two seperate filepaths or a space present in the directory naming.
Add filepathWithSpace_success to test for filepath with spaces in directory naming.