Update startPayroll and pay interactions with hoursWorked and overtime
Update startPayroll command to reset hoursWorked and overtime to 0 immediately after calculating payroll instead of having pay command reset after payment.
Update wordings in infoPanel to make it clear that amount owed is from previous payroll and the displayed hours worked and overtime is the current recorded amount since then.

Previously, the hours worked and overtime is only reset to 0 after payment, resulting in a loophole where new hours worked and overtime could still be added but is not counted towards payroll and is erased when paid.

With this update, this would fit the intended purpose of having the previous payroll being uneditable after being calculated, while still allowing the current hours worked and overtime to be updated and used for the next payroll calculation.