Fix observer implementation for InfoPanel
Note: No change in functionality in this commit, purely implementation.

Previous versions used the in-built java observable classes, which I have experienced to be quite buggy and weird, as it does not consistently update its observers.
To solve this, custom Obserable and Observer classes have been created to be used instead.